[cleanup] Some size_t changed with unsigned

+++ kernel/argument.cc
@@ -29,7 +29,7 @@ namespace vita
   /// \c i (argument(i)) varies depending on the function being evaluated
   /// (instead, adf functions have a precise, fixed signature).
   ///
-  argument::argument(size_t n)
+  argument::argument(unsigned n)
     : terminal("ARG", category_t(0), false, false, 0), index_(n)
   {
     assert(debug());
@@ -38,7 +38,7 @@ namespace vita
   ///
   /// \return the index of the argument.
   ///
-  size_t argument::index() const
+  unsigned argument::index() const
   {
     return index_;
   }
